+++ apps/gaussian_box/gaussian_box_cascaded.cpp
@@ -31,9 +31,9 @@ int main(int argc, char **argv) {
 
     // ----------------------------------------------------------------------------------------------
 
-    double sigma = 16.0;
+    float sigma = 16.0;
     int    box   = gaussian_box_filter(3, sigma); // approx Gaussian with 3 box filters
-    double norm  = std::pow(box, 3*2);            // normalizing factor
+    float norm  = std::pow(box, 3*2);            // normalizing factor
 
     // ----------------------------------------------------------------------------------------------
